Address PR #229 review: complete remaining items
Type safety:
- Create INL1/PacketWriter.h and INL1/PacketWriterRe.h with proper
  ILClass definitions and IlcaNetPacket field layout
- NetworkManager.h: Use PacketWriter/PacketWriterRe types for
  get_PacketWriter, get_PacketWriterRe, Send methods (not void*)
- Update all callers (overworld_multiplayer.cpp, team_up.cpp) to use
  auto* instead of void* for packet writer variables
- OnSessionEvent: use SessionEventData type instead of uint64_t
- MsgWindowManager: get_MsgWindow returns MsgWindow* (forward-declared)
- UIZukanRegister: Close takes UnityAction type instead of void*
- BoxWindow: remove unnecessary bool→uint8_t cast in SetTradeInfo
- StartSessionResult: define as ILStruct with errorType + bIsSuccess

Completeness fixes:
- SessionEventType: add all 10 values from dump.cs (was only 2)
- UnionWork: replace _padding[0x88] blob with all 25 named static fields
- UnionTradeManager: remove explicit padding arrays (compiler handles)
- PokemonParam copy ctor: use dst->ctor(src) instead of raw
  serialize/deserialize buffers in splitTrainerParty

+++ src/mod/externals/Dpr/UI/BoxWindow.h
@@ -264,7 +264,7 @@ namespace Dpr::UI {
 
         // Show/hide the trade info panel (partner pokemon model + info)
         inline void SetTradeInfo(bool isEnable) {
-            external<void>(0x1CB7040, this, (uint8_t)isEnable);
+            external<void>(0x1CB7040, this, isEnable);
         }
 
         // Set the partner's pokemon data for the trade info panel display
